+++ ui/js/component/app/index.js
@@ -5,6 +5,7 @@ import {
   doCheckUpgradeAvailable,
   doOpenModal,
   doAlertError,
+  doRecordScroll,
 } from "actions/app";
 import { doUpdateBalance } from "actions/wallet";
 import { selectWelcomeModalAcknowledged } from "selectors/app";
@@ -36,6 +37,7 @@ const perform = dispatch => ({
   checkUpgradeAvailable: () => dispatch(doCheckUpgradeAvailable()),
   openWelcomeModal: () => dispatch(doOpenModal(modals.WELCOME)),
   updateBalance: balance => dispatch(doUpdateBalance(balance)),
+  recordScroll: scrollPosition => dispatch(doRecordScroll(scrollPosition)),
 });
 
 export default connect(select, perform)(App);