+++ ui/js/component/publishForm/internal/channelSection.jsx
@@ -93,6 +93,7 @@ class ChannelSection extends React.PureComponent {
       "This LBC remains yours and the deposit can be undone at any time."
     );
 
+    const channel = this.state.addingChannel ? "new" : this.props.channel;
     const { fetchingChannels, channels = [] } = this.props;
 
     let channelContent = [];
@@ -102,7 +103,7 @@ class ChannelSection extends React.PureComponent {
         type="select"
         tabIndex="1"
         onChange={this.handleChannelChange.bind(this)}
-        value={this.props.channel}
+        value={channel}
       >
         <option key="anonymous" value="anonymous">
           {__("Anonymous")}
@@ -111,7 +112,7 @@ class ChannelSection extends React.PureComponent {
           <option key={name} value={name}>{name}</option>
         )}
         <option key="new" value="new">
-          {__("New identity...")}
+          {__("New channel...")}
         </option>
       </FormRow>
     );
@@ -124,9 +125,10 @@ class ChannelSection extends React.PureComponent {
     return (
       <section className="card">
         <div className="card__title-primary">
-          <h4>{__("Identity")}</h4>
+          <h4>{__("Channel Name")}</h4>
           <div className="card__subtitle">
-            {__("Who created this content?")}
+            {__("This is the channel that broadcasts your content.")}
+            {__("Ex. @Marvel, @TheBeatles, @BooksByJoe")}
           </div>
         </div>
         <div className="card__content">
@@ -137,9 +139,7 @@ class ChannelSection extends React.PureComponent {
             <FormRow
               label={__("Name")}
               type="text"
-              onChange={event => {
-                this.handleNewChannelNameChange(event);
-              }}
+              onChange={this.handleNewChannelNameChange.bind(this)}
               value={this.state.newChannelName}
             />
             <FormRow
@@ -158,8 +158,8 @@ class ChannelSection extends React.PureComponent {
                 button="primary"
                 label={
                   !this.state.creatingChannel
-                    ? __("Create identity")
-                    : __("Creating identity...")
+                    ? __("Create channel")
+                    : __("Creating channel...")
                 }
                 onClick={this.handleCreateChannelClick.bind(this)}
                 disabled={this.state.creatingChannel}