+++ ui/js/constants/modal_types.js
@@ -1,4 +1,5 @@
 export const CONFIRM_FILE_REMOVE = "confirmFileRemove";
+export const INCOMPATIBLE_DAEMON = "incompatibleDaemon";
 export const DOWNLOADING = "downloading";
 export const ERROR = "error";
 export const INSUFFICIENT_CREDITS = "insufficient_credits";